// client/src/app/networking/networkPackets/fromServer/serverPackets.ts
export enum ServerPackets {
  SPAWN_ENTITY = 'SPAWN_ENTITY',
  SPAWN_ENTITY_CHARACTER = 'SPAWN_ENTITY_CHARACTER',
  SPAWN_ENTITY_HUMANOID = 'SPAWN_ENTITY_HUMANOID',
  SPAWN_PLAYER = 'SPAWN_PLAYER',
  INIT_LEVEL = 'INIT_LEVEL',
  PLAYER_INPUT_MOVE = 'PLAYER_INPUT_MOVE',
  SYNC_SNAPSHOT = 'SYNC_SNAPSHOT',
  PONG = 'PONG',
}